막걸리 디테일 - 찜, 평가, 코멘트 정보 추가 API 수정 완료

+++ src/main/java/com/julook/domain/user/entity/Comment.java
@@ -31,7 +31,7 @@ public class Comment {
     private Long commentUserId;
 
     @Column(name = "mak_id")
-    private int commentMakId;
+    private Long commentMakId;
 
     @Column(name = "cm_content")
     private String contents;
